feat: synthMarketId collateral changes compiling

+++ contracts/modules/PerpAccountModule.sol
@@ -33,8 +33,8 @@ contract PerpAccountModule is IPerpAccountModule {
         uint256 length = globalMarginConfig.supportedSynthMarketIds.length;
         IPerpAccountModule.DepositedCollateral[] memory depositedCollaterals = new DepositedCollateral[](length);
         uint128 synthMarketId;
-        uint128 collateralAvailable;
-        uint128 collateralPrice;
+        uint256 collateralAvailable;
+        uint256 collateralPrice;
 
         for (uint256 i = 0; i < length; ) {
             synthMarketId = globalMarginConfig.supportedSynthMarketIds[i];
